+++ packages/eslint-plugin/src/util/misc.ts
@@ -133,11 +133,30 @@ function getEnumNames<T extends string>(myEnum: Record<T, unknown>): T[] {
   return Object.keys(myEnum).filter(x => isNaN(parseInt(x))) as T[];
 }
 
+/**
+ * Given an array of words, returns an English-friendly concatenation, separated with commas, with
+ * the `and` clause inserted before the last item.
+ *
+ * Example: ['foo', 'bar', 'baz' ] returns the string "foo, bar, and baz".
+ */
+function formatWordList(words: string[]): string {
+  if (!words?.length) {
+    return '';
+  }
+
+  if (words.length === 1) {
+    return words[0];
+  }
+
+  return [words.slice(0, -1).join(', '), words.slice(-1)[0]].join(' and ');
+}
+
 export {
   arraysAreEqual,
   Equal,
   ExcludeKeys,
   findFirstResult,
+  formatWordList,
   getEnumNames,
   getNameFromIndexSignature,
   getNameFromMember,